+++ test/unit/test_compiler.py
@@ -62,21 +62,11 @@ class CompilerTest(unittest.TestCase):
 
         self.config = config_from_parts_or_dicts(project_cfg, profile_cfg)
 
-        self._generate_runtime_model_patch = patch.object(dbt.compilation, 'generate_runtime_model')
+        self._generate_runtime_model_patch = patch.object(
+            dbt.compilation, 'generate_runtime_model')
         self.mock_generate_runtime_model = self._generate_runtime_model_patch.start()
 
         inject_adapter(Plugin.adapter(self.config), Plugin)
-
-        # self.mock_adapter = PostgresAdapter MagicMock(type=MagicMock(return_value='postgres'))
-        # self.mock_adapter.Relation =
-        # self.mock_adapter.get_compiler.return_value = dbt.compilation.Compiler
-        # self.mock_plugin = MagicMock(
-        #     adapter=MagicMock(
-        #         credentials=MagicMock(return_value='postgres')
-        #     )
-        # )
-        # inject_adapter(self.mock_adapter, self.mock_plugin)
-        # so we can make an adapter
 
         def mock_generate_runtime_model_context(model, config, manifest):
             def ref(name):
@@ -98,7 +88,7 @@ class CompilerTest(unittest.TestCase):
         manifest = Manifest(
             macros={},
             nodes={
-                'model.root.view': CompiledModelNode(
+                'model.root.view': ParsedModelNode(
                     name='view',
                     database='dbt',
                     schema='analytics',
@@ -108,23 +98,13 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'view'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(nodes=['model.root.ephemeral']),
                     config=self.model_config,
-                    tags=[],
                     path='view.sql',
                     original_file_path='view.sql',
-                    raw_sql='select * from {{ref("ephemeral")}}',
-                    compiled=True,
-                    extra_ctes_injected=False,
-                    extra_ctes=[InjectedCTE(id='model.root.ephemeral', sql='select * from source_table')],
-                    compiled_sql=(
-                        'with cte as (select * from something_else) '
-                        'select * from __dbt__cte__ephemeral'),
+                    raw_sql='with cte as (select * from something_else) select * from {{ref("ephemeral")}}',
                     checksum=FileHash.from_contents(''),
                 ),
-                'model.root.ephemeral': CompiledModelNode(
+                'model.root.ephemeral': ParsedModelNode(
                     name='ephemeral',
                     database='dbt',
                     schema='analytics',
@@ -134,18 +114,10 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'ephemeral'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(),
                     config=ephemeral_config,
-                    tags=[],
                     path='ephemeral.sql',
                     original_file_path='ephemeral.sql',
                     raw_sql='select * from source_table',
-                    compiled=True,
-                    compiled_sql='select * from source_table',
-                    extra_ctes_injected=False,
-                    extra_ctes=[],
                     checksum=FileHash.from_contents(''),
                 ),
             },
@@ -158,11 +130,8 @@ class CompilerTest(unittest.TestCase):
         )
 
         compiler = dbt.compilation.Compiler(self.config)
-        result, _ = compiler._recursively_prepend_ctes(
-            manifest.nodes['model.root.view'],
-            manifest,
-            {}
-        )
+        result = compiler.compile_node(
+            manifest.nodes['model.root.view'], manifest, write=False)
 
         self.assertEqual(result, manifest.nodes['model.root.view'])
         self.assertEqual(result.extra_ctes_injected, True)
@@ -172,16 +141,15 @@ class CompilerTest(unittest.TestCase):
              'select * from source_table'
              '), cte as (select * from something_else) '
              'select * from __dbt__cte__ephemeral'))
-
         self.assertEqual(
             manifest.nodes['model.root.ephemeral'].extra_ctes_injected,
-            False)
+            True)
 
     def test__prepend_ctes__no_ctes(self):
         manifest = Manifest(
             macros={},
             nodes={
-                'model.root.view': CompiledModelNode(
+                'model.root.view': ParsedModelNode(
                     name='view',
                     database='dbt',
                     schema='analytics',
@@ -191,23 +159,14 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'view'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(),
                     config=self.model_config,
-                    tags=[],
                     path='view.sql',
                     original_file_path='view.sql',
                     raw_sql=('with cte as (select * from something_else) '
                              'select * from source_table'),
-                    compiled=True,
-                    extra_ctes_injected=False,
-                    extra_ctes=[],
-                    compiled_sql=('with cte as (select * from something_else) '
-                                  'select * from source_table'),
                     checksum=FileHash.from_contents(''),
                 ),
-                'model.root.view_no_cte': CompiledModelNode(
+                'model.root.view_no_cte': ParsedModelNode(
                     name='view_no_cte',
                     database='dbt',
                     schema='analytics',
@@ -217,18 +176,10 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'view_no_cte'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(),
                     config=self.model_config,
-                    tags=[],
                     path='view.sql',
                     original_file_path='view.sql',
                     raw_sql='select * from source_table',
-                    compiled=True,
-                    extra_ctes_injected=False,
-                    extra_ctes=[],
-                    compiled_sql=('select * from source_table'),
                     checksum=FileHash.from_contents(''),
                 ),
             },
@@ -241,11 +192,8 @@ class CompilerTest(unittest.TestCase):
         )
 
         compiler = dbt.compilation.Compiler(self.config)
-        result, _ = compiler._recursively_prepend_ctes(
-            manifest.nodes['model.root.view'],
-            manifest,
-            {}
-        )
+        result = compiler.compile_node(
+            manifest.nodes['model.root.view'], manifest, write=False)
 
         self.assertEqual(
             result,
@@ -253,13 +201,13 @@ class CompilerTest(unittest.TestCase):
         self.assertTrue(result.extra_ctes_injected)
         self.assertEqualIgnoreWhitespace(
             result.compiled_sql,
-            manifest.nodes.get('model.root.view').compiled_sql)
+            ('with cte as (select * from something_else) '
+             'select * from source_table')
+        )
 
         compiler = dbt.compilation.Compiler(self.config)
-        result, _ = compiler._recursively_prepend_ctes(
-            manifest.nodes.get('model.root.view_no_cte'),
-            manifest,
-            {})
+        result = compiler.compile_node(
+            manifest.nodes['model.root.view_no_cte'], manifest, write=False)
 
         self.assertEqual(
             result,
@@ -267,7 +215,8 @@ class CompilerTest(unittest.TestCase):
         self.assertTrue(result.extra_ctes_injected)
         self.assertEqualIgnoreWhitespace(
             result.compiled_sql,
-            manifest.nodes.get('model.root.view_no_cte').compiled_sql)
+            'select * from source_table'
+        )
 
     def test__prepend_ctes(self):
         ephemeral_config = self.model_config.replace(materialized='ephemeral')
@@ -275,7 +224,7 @@ class CompilerTest(unittest.TestCase):
         manifest = Manifest(
             macros={},
             nodes={
-                'model.root.view': CompiledModelNode(
+                'model.root.view': ParsedModelNode(
                     name='view',
                     database='dbt',
                     schema='analytics',
@@ -285,21 +234,13 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'view'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(nodes=['model.root.ephemeral']),
                     config=self.model_config,
-                    tags=[],
                     path='view.sql',
                     original_file_path='view.sql',
                     raw_sql='select * from {{ref("ephemeral")}}',
-                    compiled=True,
-                    extra_ctes_injected=False,
-                    extra_ctes=[InjectedCTE(id='model.root.ephemeral', sql='select * from source_table')],
-                    compiled_sql='select * from __dbt__cte__ephemeral',
                     checksum=FileHash.from_contents(''),
                 ),
-                'model.root.ephemeral': CompiledModelNode(
+                'model.root.ephemeral': ParsedModelNode(
                     name='ephemeral',
                     database='dbt',
                     schema='analytics',
@@ -309,18 +250,10 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'ephemeral'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(),
                     config=ephemeral_config,
-                    tags=[],
                     path='ephemeral.sql',
                     original_file_path='ephemeral.sql',
                     raw_sql='select * from source_table',
-                    compiled=True,
-                    extra_ctes_injected=False,
-                    extra_ctes=[],
-                    compiled_sql='select * from source_table',
                     checksum=FileHash.from_contents(''),
                 ),
             },
@@ -333,15 +266,11 @@ class CompilerTest(unittest.TestCase):
         )
 
         compiler = dbt.compilation.Compiler(self.config)
-        result, _ = compiler._recursively_prepend_ctes(
+        result = compiler.compile_node(
             manifest.nodes['model.root.view'],
             manifest,
-            {}
+            write=False
         )
-
-        self.assertEqual(result,
-                         manifest.nodes.get('model.root.view'))
-
         self.assertTrue(result.extra_ctes_injected)
         self.assertEqualIgnoreWhitespace(
             result.compiled_sql,
@@ -349,8 +278,8 @@ class CompilerTest(unittest.TestCase):
              'select * from source_table'
              ') '
              'select * from __dbt__cte__ephemeral'))
-
-        self.assertFalse(manifest.nodes['model.root.ephemeral'].extra_ctes_injected)
+        self.assertTrue(
+            manifest.nodes['model.root.ephemeral'].extra_ctes_injected)
 
     def test__prepend_ctes__cte_not_compiled(self):
         ephemeral_config = self.model_config.replace(materialized='ephemeral')
@@ -421,7 +350,8 @@ class CompilerTest(unittest.TestCase):
                     raw_sql='select * from {{ref("ephemeral")}}',
                     compiled=True,
                     extra_ctes_injected=False,
-                    extra_ctes=[InjectedCTE(id='model.root.ephemeral', sql='select * from source_table')],
+                    extra_ctes=[InjectedCTE(
+                        id='model.root.ephemeral', sql='select * from source_table')],
                     compiled_sql='select * from __dbt__cte__ephemeral',
                     checksum=FileHash.from_contents(''),
                 ),
@@ -444,7 +374,8 @@ class CompilerTest(unittest.TestCase):
                 manifest,
                 {}
             )
-            compile_node.assert_called_once_with(parsed_ephemeral, manifest, {})
+            compile_node.assert_called_once_with(
+                parsed_ephemeral, manifest, {})
 
         self.assertEqual(result,
                          manifest.nodes.get('model.root.view'))
@@ -458,7 +389,8 @@ class CompilerTest(unittest.TestCase):
              ') '
              'select * from __dbt__cte__ephemeral'))
 
-        self.assertTrue(manifest.nodes['model.root.ephemeral'].extra_ctes_injected)
+        self.assertTrue(
+            manifest.nodes['model.root.ephemeral'].extra_ctes_injected)
 
     def test__prepend_ctes__multiple_levels(self):
         ephemeral_config = self.model_config.replace(materialized='ephemeral')
@@ -466,7 +398,7 @@ class CompilerTest(unittest.TestCase):
         manifest = Manifest(
             macros={},
             nodes={
-                'model.root.view': CompiledModelNode(
+                'model.root.view': ParsedModelNode(
                     name='view',
                     database='dbt',
                     schema='analytics',
@@ -476,18 +408,10 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'view'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(nodes=['model.root.ephemeral']),
                     config=self.model_config,
-                    tags=[],
                     path='view.sql',
                     original_file_path='view.sql',
                     raw_sql='select * from {{ref("ephemeral")}}',
-                    compiled=True,
-                    extra_ctes_injected=False,
-                    extra_ctes=[InjectedCTE(id='model.root.ephemeral', sql=None)],
-                    compiled_sql='select * from __dbt__cte__ephemeral',
                     checksum=FileHash.from_contents(''),
 
                 ),
@@ -501,11 +425,7 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'ephemeral'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(),
                     config=ephemeral_config,
-                    tags=[],
                     path='ephemeral.sql',
                     original_file_path='ephemeral.sql',
                     raw_sql='select * from {{ref("ephemeral_level_two")}}',
@@ -521,11 +441,7 @@ class CompilerTest(unittest.TestCase):
                     fqn=['root', 'ephemeral_level_two'],
                     package_name='root',
                     root_path='/usr/src/app',
-                    refs=[],
-                    sources=[],
-                    depends_on=DependsOn(),
                     config=ephemeral_config,
-                    tags=[],
                     path='ephemeral_level_two.sql',
                     original_file_path='ephemeral_level_two.sql',
                     raw_sql='select * from source_table',
@@ -541,11 +457,7 @@ class CompilerTest(unittest.TestCase):
         )
 
         compiler = dbt.compilation.Compiler(self.config)
-        result, _ = compiler._recursively_prepend_ctes(
-            manifest.nodes['model.root.view'],
-            manifest,
-            {}
-        )
+        result = compiler.compile_node(manifest.nodes['model.root.view'], manifest, write=False)
 
         self.assertEqual(result, manifest.nodes['model.root.view'])
         self.assertTrue(result.extra_ctes_injected)
@@ -559,6 +471,97 @@ class CompilerTest(unittest.TestCase):
              'select * from __dbt__cte__ephemeral'))
 
         self.assertTrue(manifest.nodes['model.root.ephemeral'].compiled)
-        self.assertTrue(manifest.nodes['model.root.ephemeral_level_two'].compiled)
-        self.assertTrue(manifest.nodes['model.root.ephemeral'].extra_ctes_injected)
-        self.assertTrue(manifest.nodes['model.root.ephemeral_level_two'].extra_ctes_injected)
+        self.assertTrue(
+            manifest.nodes['model.root.ephemeral_level_two'].compiled)
+        self.assertTrue(
+            manifest.nodes['model.root.ephemeral'].extra_ctes_injected)
+        self.assertTrue(
+            manifest.nodes['model.root.ephemeral_level_two'].extra_ctes_injected)
+
+    def test__prepend_ctes__valid_ephemeral_sql(self):
+        """Assert that the compiled sql for ephemeral models is valid and can be executed on its own"""
+        ephemeral_config = self.model_config.replace(materialized='ephemeral')
+
+        manifest = Manifest(
+            macros={},
+            nodes={
+                'model.root.view': ParsedModelNode(
+                    name='view',
+                    database='dbt',
+                    schema='analytics',
+                    alias='view',
+                    resource_type=NodeType.Model,
+                    unique_id='model.root.view',
+                    fqn=['root', 'view'],
+                    package_name='root',
+                    root_path='/usr/src/app',
+                    config=self.model_config,
+                    path='view.sql',
+                    original_file_path='view.sql',
+                    raw_sql='select * from {{ref("ephemeral")}}',
+                    checksum=FileHash.from_contents(''),
+                ),
+                'model.root.inner_ephemeral': ParsedModelNode(
+                    name='inner_ephemeral',
+                    database='dbt',
+                    schema='analytics',
+                    alias='inner_ephemeral',
+                    resource_type=NodeType.Model,
+                    unique_id='model.root.inner_ephemeral',
+                    fqn=['root', 'inner_ephemeral'],
+                    package_name='root',
+                    root_path='/usr/src/app',
+                    config=ephemeral_config,
+                    path='inner_ephemeral.sql',
+                    original_file_path='inner_ephemeral.sql',
+                    raw_sql='select * from source_table',
+                    checksum=FileHash.from_contents(''),
+                ),
+                 'model.root.ephemeral': ParsedModelNode(
+                    name='ephemeral',
+                    database='dbt',
+                    schema='analytics',
+                    alias='ephemeral',
+                    resource_type=NodeType.Model,
+                    unique_id='model.root.ephemeral',
+                    fqn=['root', 'ephemeral'],
+                    package_name='root',
+                    root_path='/usr/src/app',
+                    config=ephemeral_config,
+                    path='ephemeral.sql',
+                    original_file_path='ephemeral.sql',
+                    raw_sql='select * from {{ ref("inner_ephemeral") }}',
+                    checksum=FileHash.from_contents(''),
+                ),
+            },
+            sources={},
+            docs={},
+            disabled=[],
+            files={},
+            exposures={},
+            selectors={},
+        )
+
+        compiler = dbt.compilation.Compiler(self.config)
+        result = compiler.compile_node(
+            manifest.nodes['model.root.view'],
+            manifest,
+            write=False
+        )
+        self.assertEqualIgnoreWhitespace(
+            result.compiled_sql,
+            ('with __dbt__cte__inner_ephemeral as ('
+             'select * from source_table'
+             '), '
+             '__dbt__cte__ephemeral as ('
+             'select * from __dbt__cte__inner_ephemeral'
+             ') '
+             'select * from __dbt__cte__ephemeral'))
+        self.assertEqualIgnoreWhitespace(
+            manifest.nodes['model.root.ephemeral'].compiled_sql,
+            ('with __dbt__cte__inner_ephemeral as ('
+            'select * from source_table'
+            ')  '
+            'select * from __dbt__cte__inner_ephemeral')
+        )
+