+++ CHANGELOG.md
@@ -92,6 +92,10 @@ Contributors:
 
 - Fix compiled sql for ephemeral models ([#3139](https://github.com/fishtown-analytics/dbt/pull/3139), [#3056](https://github.com/fishtown-analytics/dbt/pull/3056))
 
+### Fixes
+
+- Fix compiled sql for ephemeral models ([#3139](https://github.com/fishtown-analytics/dbt/pull/3139), [#3056](https://github.com/fishtown-analytics/dbt/pull/3056))
+
 ## dbt 0.19.1b2 (February 15, 2021)
 
 ## dbt 0.19.1b1 (February 12, 2021)